+++ lib/packages/wikibase/data-model/src/Term/AbstractTerm.php
@@ -1,0 +1,39 @@
+<?php
+
+namespace Wikibase\DataModel\Term;
+
+/**
+ * @since 0.7.3
+ *
+ * @licence GNU GPL v2+
+ * @author Jeroen De Dauw < jeroendedauw@gmail.com >
+ */
+abstract class AbstractTerm implements Term {
+
+	private $languageCode;
+	private $text;
+
+	/**
+	 * @param string $languageCode
+	 * @param string $text
+	 */
+	public function __construct( $languageCode, $text ) {
+		$this->languageCode = $languageCode;
+		$this->text = $text;
+	}
+
+	/**
+	 * @return string
+	 */
+	public function getLanguageCode() {
+		return $this->languageCode;
+	}
+
+	/**
+	 * @return string
+	 */
+	public function getText() {
+		return $this->text;
+	}
+
+}