+++ lib/packages/wikibase/data-model/src/Term/FingerprintProvider.php
@@ -1,0 +1,18 @@
+<?php
+
+namespace Wikibase\DataModel\Term;
+
+/**
+ * @since 0.7.3
+ *
+ * @licence GNU GPL v2+
+ * @author Jeroen De Dauw < jeroendedauw@gmail.com >
+ */
+interface FingerprintProvider {
+
+	/**
+	 * @return Fingerprint
+	 */
+	public function getFingerprint();
+
+}